+++ NEWS.md
@@ -23,6 +23,7 @@
 
 ## new
 - `sliceGiotto()` for pulling out specific spatial units and feature types as independent `giotto` objects
+- `splitGiotto()` for splitting a Giotto object into a list of Giotto objects based on a cell metadata column
 - `as.list()` method for `giotto` to dump the data as a list of subobjects