+++ NEWS.md
@@ -3,6 +3,7 @@
 
 ## breaking changes
 - stop exporting deprecated internal accessors
+- terra requirement raised to 1.7.41 for `minCircle()`
 
 ## bug fixes
 - fix `dimnames()` for some subobjects
@@ -34,6 +35,7 @@
 - `splitGiotto()` for splitting a Giotto object into a list of Giotto objects based on a cell metadata column
 - `as.list()` method for `giotto` to dump the data as a list of subobjects
 - `XY()` and `XY<-()` for accessing and setting coordinate values of subobjects as `matrix`
+- terra `convHull()`, `minRect()`, `minCircle()` for Giotto spatial vector classes
 
 
 # GiottoClass 0.3.5 (2024/08/28)